+++ backend/src/core/interfaces/campaign.interface.ts
@@ -49,6 +49,7 @@ export interface CampaignDetails {
 
 export interface CampaignStats extends CampaignStatsCount {
   status: string
+  updatedAt: Date
 }
 
 export interface CampaignStatsCount {
@@ -56,16 +57,12 @@ export interface CampaignStatsCount {
   unsent: number
   sent: number
   invalid: number
-  status?: string
-  updated_at?: Date
 }
 
 export interface CampaignInvalidRecipient {
   recipient: string
   messageId: string
   errorCode: string
-  sentAt: Date
-  updatedAt: Date
 }
 
 /**
