+++ frontend/src/components/common/failed-recipients-csv/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FailedRecipientsCsv'
