+++ frontend/src/components/common/index.ts
@@ -22,4 +22,4 @@ export { default as Dropdown } from './dropdown'
 export { default as BodyWrapper } from './body-wrapper'
 export { default as CsvUpload } from './csv-upload'
 export { default as InfoBanner } from './info-banner'
-export { default as InvalidRecipientsCsv } from './invalid-recipients-csv'
+export { default as FailedRecipientsCsv } from './failed-recipients-csv'
