--- frontend/src/components/common/invalid-recipients-csv/index.ts
@@ -1,1 +1,0 @@
-export { default } from './InvalidRecipientsCsv'
